# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sala Situacional - Sistema Colaborativo de Alertas",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://10.10.80.70:3090",
      "description": "Servidor Principal (10.10.80.70:3090)"
    },
    {
      "url": "http://localhost:3090",
      "description": "Servidor Local (localhost:3090)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido tras el inicio de sesión (/auth/login)"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-token",
        "description": "Token opcional de sistema"
      }
    },
    "schemas": {
...
      "CrearReporte": {
        "type": "object",
        "properties": {
          "titulo": {
            "type": "string",
            "example": "Reporte de Novedades e Incidentes - Turno Mañana"
          },
          "observaciones_generales": {
            "type": "string",
            "example": "Monitoreo en tiempo real de lluvias y acumulación de agua."
          },
          "numero_rds": {
            "type": "string",
            "example": "SEGURA-EP-GASGEC-SS-2026-041 (Lluvias)"
          },
          "fecha_reporte": {
            "type": "string",
            "example": "2026-08-25"
          },
          "hora_inicio": {
            "type": "string",
            "example": "06:00"
          },
          "hora_fin": {
            "type": "string",
            "example": "22:00"
          },
          "revisado_por": {
            "type": "string",
            "example": "Jefe de Sala Situacional | MSc. Ing. Santiago Jaramillo"
          },
          "cabecera": {
            "type": "string",
            "example": "REPORTE DE NOVEDADES POR LLUVIAS INICIAL: 07/05/2026 21h30"
          },
          "periodo": {
            "type": "string",
            "example": "Durante la noche del 7 de mayo se han registrado las siguientes novedades en el cantón Guayaquil por efecto de las lluvias:"
          },
          "inocar_fecha": {
            "type": "string",
            "example": "7 de mayo"
          },
          "inocar_pleamar": {
            "type": "string",
            "example": "a las 22h42 con 4.13m"
          },
          "inocar_bajamar": {
            "type": "string",
            "example": "a las 05h27 del 08/05/2026 con 0.79m"
          },
          "usuario_id": {
            "type": "integer",
            "example": 1,
            "description": "ID de usuario para agregarlo como colaborador inicial"
          },
          "correo_colaborador": {
            "type": "string",
            "example": "operador2@segura.gob.ec",
            "description": "Correo del usuario a agregar como colaborador inicial"
          },
          "colaboradores": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "operador2@segura.gob.ec",
              "operador3@segura.gob.ec"
            ],
            "description": "Lista de IDs o correos de colaboradores a vincular"
          }
        }
      },
      "ActualizarParametrosReporte": {
        "type": "object",
        "properties": {
          "numero_rds": {
            "type": "string",
            "example": "SEGURA-EP-GASGEC-SS-2026-041 (Lluvias)"
          },
          "fecha_reporte": {
            "type": "string",
            "example": "2026-08-25"
          },
          "hora_inicio": {
            "type": "string",
            "example": "06:00"
          },
          "hora_fin": {
            "type": "string",
            "example": "22:00"
          },
          "revisado_por": {
            "type": "string",
            "example": "Jefe de Sala Situacional | MSc. Ing. Santiago Jaramillo"
          },
          "cabecera": {
            "type": "string",
            "example": "REPORTE DE NOVEDADES POR LLUVIAS INICIAL: 07/05/2026 21h30"
          },
          "periodo": {
            "type": "string",
            "example": "Durante la noche del 7 de mayo se han registrado las siguientes novedades en el cantón Guayaquil por efecto de las lluvias:"
          },
          "inocar_fecha": {
            "type": "string",
            "example": "7 de mayo"
          },
          "inocar_pleamar": {
            "type": "string",
            "example": "a las 22h42 con 4.13m"
          },
          "inocar_bajamar": {
            "type": "string",
            "example": "a las 05h27 del 08/05/2026 con 0.79m"
          },
          "observaciones_generales": {
            "type": "string",
            "example": "Observaciones actualizadas."
          }
        }
      },
      "AgregarNovedad": {
        "type": "object",
        "required": [
          "direccion"
        ],
        "properties": {
          "tipo_evento": {
            "type": "string",
            "enum": [
              "AGUA",
              "ARBOL",
              "DESLIZAMIENTO",
              "POSTE",
              "SINIESTRO",
              "INUNDACION",
              "VENDAVAL",
              "AFECTACION"
            ],
            "example": "AGUA"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "AGUA",
              "ARBOL",
              "DESLIZAMIENTO",
              "POSTE",
              "SINIESTRO",
              "INUNDACION",
              "VENDAVAL",
              "AFECTACION"
            ],
            "example": "AGUA"
          },
          "direccion": {
            "type": "string",
            "example": "PROSPERINA 6TO CALLEJON Y AV 41 DIAGONAL A LAS ROSAS"
          },
          "aga": {
            "type": "string",
            "example": "A09"
          },
          "instituciones": {
            "type": "string",
            "example": "@emapagye @interagua"
          },
          "fecha_evento": {
            "type": "string",
            "example": "2026-08-25"
          },
          "hora_evento": {
            "type": "string",
            "example": "16:40"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-25T16:40:00Z"
          },
          "latitud": {
            "type": "number",
            "example": -2.1894
          },
          "longitud": {
            "type": "number",
            "example": -79.8891
          },
          "recurso": {
            "type": "string",
            "example": "INS-ALC"
          },
          "recurso_asignado": {
            "type": "string",
            "example": "INS-ALC"
          },
          "estado": {
            "type": "string",
            "enum": [
              "PENDIENTE",
              "EN_SITIO",
              "EN_ATENCION",
              "SOLUCIONADO"
            ],
            "example": "PENDIENTE"
          },
          "estado_operativo": {
            "type": "string",
            "example": "PENDIENTE"
          },
          "descripcion": {
            "type": "string",
            "example": "Acumulación de agua considerable en calzada con afectación al tránsito."
          },
          "acciones": {
            "type": "string",
            "example": "Se despachó cuadrilla de Interagua."
          },
          "acciones_inmediatas": {
            "type": "string",
            "example": "Se despachó cuadrilla de Interagua."
          },
          "hora_sitio": {
            "type": "string",
            "example": "17:10"
          },
          "solucionado": {
            "type": "string",
            "example": "18:30"
          },
          "reporte_id": {
            "type": "integer",
            "example": 1,
            "description": "ID opcional del reporte al que se asocia la novedad"
          },
          "datos_adicionales": {
            "type": "object",
            "description": "Metadatos adicionales en formato JSONB (ficha, camaras, conteo de recursos, etc.)",
            "example": {
              "ficha": "FICHA-2026-001",
              "camara_cvvc": "CAM-09",
              "via_afectada": "SI",
              "recursos": {
                "BCBG": 1,
                "ATM": 2
              },
              "personal": {
                "BCBG": 3,
                "ATM": 4
              }
            }
          },
          "recursos_instituciones": {
            "type": "object",
            "description": "Desglose de vehículos/recursos por institución",
            "example": {
              "BCBG": 1,
              "ATM": 2
            }
          },
          "personal_instituciones": {
            "type": "object",
            "description": "Desglose de personal/efectivos por institución",
            "example": {
              "BCBG": 3,
              "ATM": 4
            }
          },
          "fotos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "/uploads/fotos/foto-1724678123-123456789.jpg"
            ]
          }
        }
      },
      "ActualizarNovedad": {
        "type": "object",
        "properties": {
          "tipo_evento": {
            "type": "string",
            "enum": [
              "AGUA",
              "ARBOL",
              "DESLIZAMIENTO",
              "POSTE",
              "SINIESTRO",
              "INUNDACION",
              "VENDAVAL",
              "AFECTACION"
            ],
            "example": "AGUA"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "AGUA",
              "ARBOL",
              "DESLIZAMIENTO",
              "POSTE",
              "SINIESTRO",
              "INUNDACION",
              "VENDAVAL",
              "AFECTACION"
            ],
            "example": "AGUA"
          },
          "direccion": {
            "type": "string",
            "example": "PROSPERINA 6TO CALLEJON Y AV 41 DIAGONAL A LAS ROSAS"
          },
          "aga": {
            "type": "string",
            "example": "A09"
          },
          "instituciones": {
            "type": "string",
            "example": "@emapagye @interagua"
          },
          "fecha_evento": {
            "type": "string",
            "example": "2026-08-25"
          },
          "hora_evento": {
            "type": "string",
            "example": "16:40"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "example": "2026-08-25T16:40:00Z"
          },
          "latitud": {
            "type": "number",
            "example": -2.1894
          },
          "longitud": {
            "type": "number",
            "example": -79.8891
          },
          "recurso": {
            "type": "string",
            "example": "INS-ALC"
          },
          "recurso_asignado": {
            "type": "string",
            "example": "INS-ALC"
          },
          "estado": {
            "type": "string",
            "enum": [
              "PENDIENTE",
              "EN_SITIO",
              "EN_ATENCION",
              "SOLUCIONADO"
            ],
            "example": "SOLUCIONADO"
          },
          "estado_operativo": {
            "type": "string",
            "example": "SOLUCIONADO"
          },
          "descripcion": {
            "type": "string",
            "example": "Novedad atendida y vía habilitada."
          },
          "acciones": {
            "type": "string",
            "example": "Se realizó limpieza de sumideros."
          },
          "acciones_inmediatas": {
            "type": "string",
            "example": "Se realizó limpieza de sumideros."
          },
          "hora_sitio": {
            "type": "string",
            "example": "17:10"
          },
          "solucionado": {
            "type": "string",
            "example": "18:30"
          },
          "reporte_id": {
            "type": "integer",
            "example": 1
          },
          "datos_adicionales": {
            "type": "object",
            "example": {
              "ficha": "FICHA-2026-001",
              "solucionado": "SI",
              "recursos": {
                "BCBG": 1,
                "ATM": 2
              },
              "personal": {
                "BCBG": 3,
                "ATM": 4
              }
            }
          },
          "recursos_instituciones": {
            "type": "object",
            "example": {
              "BCBG": 1,
              "ATM": 2
            }
          },
          "personal_instituciones": {
            "type": "object",
            "example": {
              "BCBG": 3,
              "ATM": 4
            }
          },
          "fotos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "/uploads/fotos/foto-1724678123-123456789.jpg"
            ]
          }
        }
      }
    }
  },
  "paths": {
...
    "/reportes": {
      "get": {
        "summary": "Listar reportes (Paginación opcional, filtros de búsqueda y fechas)",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Número de página (opcional, ej. 1)"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Cantidad de registros por página (opcional, ej. 15)"
          },
          {
            "in": "query",
            "name": "busqueda",
            "schema": {
              "type": "string"
            },
            "description": "Búsqueda por número RDS, título o elaborado por"
          },
          {
            "in": "query",
            "name": "fechaDesde",
            "schema": {
              "type": "string"
            },
            "description": "Fecha inicio (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "fechaHasta",
            "schema": {
              "type": "string"
            },
            "description": "Fecha fin (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista o página de reportes"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo reporte",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearReporte"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reporte creado exitosamente"
          }
        }
      }
    },
    "/reportes/upload-foto": {
      "post": {
        "summary": "Subir fotografías de novedad al servidor local (Máx. 2 imágenes, 5MB c/u)",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "fotos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fotos subidas y guardadas exitosamente en el servidor"
          },
          "400": {
            "description": "No se enviaron archivos o formato no permitido"
          }
        }
      }
    },
    "/reportes/{id}": {
      "get": {
        "summary": "Obtener reporte por ID con sus novedades y colaboradores",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte"
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte encontrado"
          },
          "404": {
            "description": "Reporte no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un reporte completo por ID",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte eliminado exitosamente"
          },
          "400": {
            "description": "ID de reporte inválido"
          },
          "404": {
            "description": "Reporte no encontrado"
          }
        }
      }
    },
    "/reportes/{id}/parametros": {
      "put": {
        "summary": "Actualizar parámetros institucionales del reporte (Sección 2 - RDS e INOCAR)",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarParametrosReporte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parámetros del reporte actualizados exitosamente"
          },
          "404": {
            "description": "Reporte no encontrado"
          }
        }
      }
    },
    "/reportes/{id}/novedades": {
      "post": {
        "summary": "Agregar una novedad al reporte (1:N) vinculada directamente al Usuario con fotos",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgregarNovedad"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipo_evento": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "aga": {
                    "type": "string"
                  },
                  "instituciones": {
                    "type": "string"
                  },
                  "fecha_evento": {
                    "type": "string"
                  },
                  "hora_evento": {
                    "type": "string"
                  },
                  "latitud": {
                    "type": "number"
                  },
                  "longitud": {
                    "type": "number"
                  },
                  "recurso_asignado": {
                    "type": "string"
                  },
                  "estado_operativo": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "acciones_inmediatas": {
                    "type": "string"
                  },
                  "fotos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Novedad agregada y campo 'Elaborado por' actualizado"
          }
        }
      }
    },
    "/reportes/{id}/novedades/{novedadId}": {
      "put": {
        "summary": "Actualizar una novedad existente del reporte",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte"
          },
          {
            "in": "path",
            "name": "novedadId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico de la novedad a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarNovedad"
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "tipo_evento": {
                    "type": "string"
                  },
                  "direccion": {
                    "type": "string"
                  },
                  "aga": {
                    "type": "string"
                  },
                  "instituciones": {
                    "type": "string"
                  },
                  "fecha_evento": {
                    "type": "string"
                  },
                  "hora_evento": {
                    "type": "string"
                  },
                  "latitud": {
                    "type": "number"
                  },
                  "longitud": {
                    "type": "number"
                  },
                  "recurso_asignado": {
                    "type": "string"
                  },
                  "estado_operativo": {
                    "type": "string"
                  },
                  "descripcion": {
                    "type": "string"
                  },
                  "acciones_inmediatas": {
                    "type": "string"
                  },
                  "fotos": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Novedad actualizada exitosamente"
          },
          "404": {
            "description": "Reporte o novedad no encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una novedad del reporte",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte"
          },
          {
            "in": "path",
            "name": "novedadId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico de la novedad a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Novedad eliminada exitosamente"
          },
          "404": {
            "description": "Reporte o novedad no encontrada"
          }
        }
      }
    },
    "/reportes/{id}/metricas-tiempos": {
      "get": {
        "summary": "Obtener analítica de tiempos de respuesta y personal por institución para este reporte",
        "tags": [
          "Reportes"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID numérico del reporte"
          }
        ],
        "responses": {
          "200": {
            "description": "Métricas agregadas de tiempos de respuesta del reporte"
          },
          "404": {
            "description": "Reporte no encontrado"
          }
        }
      }
    },
...
  },
...
}